Implemented DAO interface, some refactorings

+++ EmployeeManagement - Jackson/src/main/java/com/employee/services/DeleteEmployee.java
@@ -12,6 +12,7 @@ import org.json.simple.parser.ParseException;
 
 import com.employee.dao.EmployeeDAO;
 import com.employee.dao.EmployeeDAOImpl;
+import com.employee.dao.ServerSideValidations;
 import com.employee.exception.EmployeeDoesNotExistException;
 import com.employee.util.EmployeeUtil;
 
@@ -19,13 +20,15 @@ public class DeleteEmployee {
 	public void delete() {
 		EmployeeDAO dao = new EmployeeDAOImpl();
 		EmployeeUtil util = new EmployeeUtil();
+		ServerSideValidations validations = new ServerSideValidations();
 		GetEmployee getEmployee = new GetEmployee();
 		Scanner sc = new Scanner(System.in);
 
 		System.out.println("Enter empId to delete:");
 		String delId = sc.next();
-
-		boolean present = util.checkEmployee(delId); // Check employee
+		if(!util.validateID(delId)) return;
+		
+		boolean present = validations.checkEmployee(delId); // Check employee
 		if (present) {
 			dao.deleteEmployee(delId);
 			getEmployee.get_all(); // SHOW records after every operation
